+++ packages/core/src/actions/updateSession/updateSession.ts
@@ -3,12 +3,16 @@ import { createEndpoint, createEndpointConfig } from "@aura-stack/router"
 import { updateSession } from "@/api/updateSession.ts"
 import type { User } from "@/@types/session.ts"
 import type { IdentityConfig } from "@/@types/config.ts"
+import { UserIdentity } from "@/shared/identity.ts"
 
-export const config = (identity: IdentityConfig) => {
+export const config = (_identity: IdentityConfig) => {
     return createEndpointConfig({
         schemas: {
             body: z.object({
-                user: identity.schema?.partial().optional(),
+                /**
+                 * @todo add support for valibot schemas in the body as well, currently only Zod is supported
+                 */
+                user: UserIdentity.partial().optional(),
                 expires: z.coerce.date().optional(),
             }),
         },
@@ -24,8 +28,8 @@ export const updateSessionAction = (identity: IdentityConfig) => {
                 ctx: ctx.context,
                 headers: ctx.request.headers,
                 session: {
-                    user: ctx.body.user as User,
-                    expires: ctx.body.expires?.toISOString(),
+                    user: ctx.body?.user as User,
+                    expires: ctx.body?.expires?.toISOString(),
                 },
             })
             return toResponse()