+++ packages/core/test/types.test-d.ts
@@ -1,7 +1,8 @@
 import { describe, expectTypeOf } from "vitest"
-import { z, ZodOptional, ZodString } from "zod/v4"
+import * as valibot from "valibot"
 import { createAuth } from "@/createAuth.ts"
-import { UserIdentity } from "@/shared/identity.ts"
+import { z, ZodOptional, ZodString } from "zod/v4"
+import { Identities, UserIdentity, UserIdentityValibot, UserShapeValibot } from "@/shared/identity.ts"
 import { github, type GitHubProfile } from "@/oauth/github.ts"
 import type {
     GetSessionAPIOptions,
@@ -13,12 +14,12 @@ import type {
 } from "@/@types/index.ts"
 import type { AuthConfig, AuthInstance, User } from "@/index.ts"
 import type { OAuthProviderCredentials } from "@/@types/oauth.ts"
-import type { EditableShape, InferSession, InferUser, ZodShapeToObject } from "@/@types/utility.ts"
+import type { EditableShape, FromShapeToObject, InferSession, InferUser, ValibotShapeToObject, ZodShapeToObject } from "@/@types/utility.ts"
 import type { JWTHeaderParameters, JWTVerifyOptions, Prettify, TypedJWTPayload } from "@aura-stack/jose"
 
 describe("createAuth", () => {
     expectTypeOf(createAuth).toEqualTypeOf<
-        <Identity extends EditableShape<UserShape>>(config: AuthConfig<Identity>) => AuthInstance<ZodShapeToObject<Identity>>
+        <Identity extends Identities = EditableShape<UserShape>>(config: AuthConfig<Identity>) => AuthInstance<FromShapeToObject<Identity>>
     >()
     expectTypeOf(createAuth({ oauth: [] }).api.getSession).toEqualTypeOf<
         (options: GetSessionAPIOptions) => Promise<GetSessionAPIReturn<ZodShapeToObject<UserShape>>>
@@ -28,7 +29,7 @@ describe("createAuth", () => {
     >()
 
     expectTypeOf(createAuth({ oauth: [] }).jose.signJWS).toEqualTypeOf<
-        (payload: TypedJWTPayload<Partial<User>>, options?: JWTHeaderParameters) => Promise<string>
+        (payload: TypedJWTPayload<Partial<ZodShapeToObject<UserShape>>>, options?: JWTHeaderParameters) => Promise<string>
     >()
     expectTypeOf(createAuth({ oauth: [] }).jose.verifyJWS).toEqualTypeOf<
         (token: string, options?: JWTVerifyOptions) => Promise<TypedJWTPayload<ZodShapeToObject<UserShape>>>
@@ -58,15 +59,76 @@ describe("createAuth", () => {
         ) => Promise<string>
     >()
     expectTypeOf(
+        createAuth({
+            oauth: [],
+            identity: { schema: valibot.object({ ...UserIdentityValibot.entries, role: valibot.string() }) },
+        }).jose.signJWS
+    ).toEqualTypeOf<
+        (
+            payload: TypedJWTPayload<
+                Partial<
+                    {
+                        sub: string
+                        role: string
+                        name?: string | null | undefined
+                        image?: string | null | undefined
+                        email?: string | null | undefined
+                    } & {
+                        sub: string
+                        name?: string | null | undefined
+                        image?: string | null | undefined
+                        email?: string | null | undefined
+                    }
+                >
+            >,
+            options?: JWTHeaderParameters
+        ) => Promise<string>
+    >()
+    expectTypeOf(
         createAuth({ oauth: [], identity: { schema: UserIdentity.extend({ role: z.string() }) } }).jose.verifyJWS
     ).toEqualTypeOf<
         (token: string, options?: JWTVerifyOptions) => Promise<TypedJWTPayload<ZodShapeToObject<UserShape & { role: ZodString }>>>
     >()
+    expectTypeOf(
+        createAuth({
+            oauth: [],
+            identity: { schema: valibot.object({ ...UserIdentityValibot.entries, role: valibot.string() }) },
+        }).jose.verifyJWS
+    ).toEqualTypeOf<
+        (
+            token: string,
+            options?: JWTVerifyOptions
+        ) => Promise<TypedJWTPayload<ValibotShapeToObject<UserShapeValibot & { role: valibot.StringSchema<undefined> }>>>
+    >()
+    expectTypeOf(createAuth({ oauth: [], identity: { schema: UserIdentityValibot } })).toEqualTypeOf<
+        AuthInstance<ValibotShapeToObject<UserShapeValibot>>
+    >()
+    expectTypeOf(
+        createAuth({
+            oauth: [],
+            identity: {
+                schema: valibot.object({
+                    ...UserIdentityValibot.entries,
+                    role: valibot.string(),
+                }),
+            },
+        })
+    ).toEqualTypeOf<AuthInstance<User & { role: string }>>()
 
     expectTypeOf(
         createAuth({ oauth: [], identity: { schema: UserIdentity.extend({ role: z.string() }) } }).api.getSession
     ).toEqualTypeOf<
         (options: GetSessionAPIOptions) => Promise<GetSessionAPIReturn<ZodShapeToObject<UserShape & { role: ZodString }>>>
+    >()
+    expectTypeOf(
+        createAuth({
+            oauth: [],
+            identity: { schema: valibot.object({ ...UserIdentityValibot.entries, role: valibot.string() }) },
+        }).api.getSession
+    ).toEqualTypeOf<
+        (
+            options: GetSessionAPIOptions
+        ) => Promise<GetSessionAPIReturn<ValibotShapeToObject<UserShapeValibot & { role: valibot.StringSchema<undefined> }>>>
     >()
     expectTypeOf(
         createAuth({ oauth: [], identity: { schema: UserIdentity.extend({ role: z.string() }) } }).api.updateSession
@@ -75,14 +137,39 @@ describe("createAuth", () => {
             options: UpdateSessionAPIOptions<ZodShapeToObject<UserShape & { role: ZodString }>>
         ) => Promise<UpdateSessionAPIReturn<ZodShapeToObject<UserShape & { role: ZodString }>>>
     >()
+    expectTypeOf(
+        createAuth({
+            oauth: [],
+            identity: { schema: valibot.object({ ...UserIdentityValibot.entries, role: valibot.string() }) },
+        }).api.updateSession
+    ).toEqualTypeOf<
+        (
+            options: UpdateSessionAPIOptions<ValibotShapeToObject<UserShapeValibot & { role: valibot.StringSchema<undefined> }>>
+        ) => Promise<UpdateSessionAPIReturn<ValibotShapeToObject<UserShapeValibot & { role: valibot.StringSchema<undefined> }>>>
+    >()
+
     expectTypeOf<InferUser<ReturnType<typeof createAuth>>>().toEqualTypeOf<User>()
-    expectTypeOf<InferSession<ReturnType<typeof createAuth>>>().toEqualTypeOf<Session<User>>()
     expectTypeOf<InferUser<ReturnType<typeof createAuth<UserShape & { role: ZodOptional<ZodString> }>>>>().toEqualTypeOf<
         Prettify<User & { role?: string | undefined }>
     >()
+    expectTypeOf<
+        InferUser<
+            ReturnType<
+                typeof createAuth<UserShapeValibot & { role: valibot.OptionalSchema<valibot.StringSchema<undefined>, undefined> }>
+            >
+        >
+    >().toEqualTypeOf<Prettify<User & { role?: string | undefined }>>()
+
     expectTypeOf<InferSession<ReturnType<typeof createAuth<UserShape & { role: ZodOptional<ZodString> }>>>>().toEqualTypeOf<
         Session<Prettify<User & { role?: string | undefined }>>
     >()
+    expectTypeOf<
+        InferSession<
+            ReturnType<
+                typeof createAuth<UserShapeValibot & { role: valibot.OptionalSchema<valibot.StringSchema<undefined>, undefined> }>
+            >
+        >
+    >().toEqualTypeOf<Session<Prettify<User & { role?: string | undefined }>>>()
 })
 
 describe("OAuth providers", () => {